refactor: 폴더 구조 변경 및 Storybook Title경로 수정
