+++ src/common/components/Icon/type.ts
