+++ src/common/components/Image/Image.variants.ts
